refactor(tds-1210): make daikon-kafka-mongo a storage implementation of daikon-kafka
