+++ daikon-kafka-mongo/src/main/java/org/talend/daikon/mongo/DaikonMongoConfiguration.java
@@ -1,0 +1,11 @@
+package org.talend.daikon.mongo;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@ComponentScan(basePackages = { "org.talend.daikon.mongo" })
+@Configuration
+@EnableAutoConfiguration
+public class DaikonMongoConfiguration {
+}
