+++ daikon-kafka-mongo/src/main/java/org/talend/daikon/mongo/util/MongoCustomConverter.java
