+++ daikon-kafka-mongo/src/main/java/org/talend/daikon/mongo/util/RepoConfiguration.java
