+++ daikon-kafka-mongo/src/test/java/org/talend/daikon/mongo/util/TestRepositoryAbstract.java
