+++ daikon-kafka-mongo/src/test/resources/logback.xml
