+++ daikon-kafka/src/main/java/org/talend/daikon/DaikonKafkaConfiguration.java
@@ -6,7 +6,6 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-@SpringBootApplication
 @ComponentScan(basePackages = { "org.talend.daikon.kafka" })
 @Configuration
 @EnableScheduling
