--- daikon-kafka/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-daikon.kafka.record.sending.enable=true
