--- daikon-kafka/src/test/resources/application.properties
@@ -1,2 +1,0 @@
-daikon.kafka.record.sending.application=testApp
-daikon.kafka.record.sending.enable=false
