+++ daikon-kafka/src/test/resources/log4j.properties
@@ -1,0 +1,23 @@
+#
+# ============================================================================
+#
+# Copyright (C) 2006-2015 Talend Inc. - www.talend.com
+#
+# This source code is available under agreement available at
+# %InstallDIR%\features\org.talend.rcp.branding.%PRODUCTNAME%\%PRODUCTNAME%license.txt
+#
+# You should have received a copy of the agreement
+# along with this program; if not, write to Talend SA
+# 9 rue Pages 92150 Suresnes, France
+#
+# ============================================================================
+#
+
+# Root logger option
+log4j.rootLogger=WARN, stdout
+
+# Direct log messages to stdout
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.Target=System.out
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
