--- daikon-mongo/src/main/java/org/talend/daikon/mongo/DaikonMongoConfiguration.java
@@ -1,13 +1,0 @@
-package org.talend.daikon.mongo;
-
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@SpringBootApplication
-@ComponentScan(basePackages = { "org.talend.daikon.mongo" })
-@Configuration
-@EnableAutoConfiguration
-public class DaikonMongoConfiguration {
-}
