--- daikon-mongo/src/main/resources/application.properties
@@ -1,4 +1,0 @@
-# Mongo database location
-spring.data.mongodb.host=localhost
-spring.data.mongodb.port=27017
-spring.data.mongodb.database=records
