+++ pom.xml
@@ -28,8 +28,8 @@
         <module>reporting</module>
         <module>daikon-content-service</module>
 		<module>daikon-logging</module>
-	    <module>daikon-mongo</module>
-	    <module>daikon-kafka</module>
+        <module>daikon-kafka</module>
+        <module>daikon-kafka-mongo</module>
     </modules>
     <scm>
         <connection>scm:git:https://github.com/Talend/daikon.git</connection>
